+++ runtime/wasmhost/wasmhost.go
@@ -19,6 +19,7 @@ import (
 	"github.com/hypermodeinc/modus/runtime/logger"
 	"github.com/hypermodeinc/modus/runtime/middleware"
 	"github.com/hypermodeinc/modus/runtime/plugins"
+	"github.com/hypermodeinc/modus/runtime/timezones"
 	"github.com/hypermodeinc/modus/runtime/utils"
 
 	"github.com/rs/zerolog"
@@ -111,6 +112,14 @@ func (host *wasmHost) GetModuleInstance(ctx context.Context, plugin *plugins.Plu
 	wOut := io.MultiWriter(buffers.StdOut(), wInfoLog)
 	wErr := io.MultiWriter(buffers.StdErr(), wErrorLog)
 
+	// Get the time zone to pass to the module instance.
+	var timeZone string
+	if tz, ok := ctx.Value(utils.TimeZoneContextKey).(string); ok {
+		timeZone = tz
+	} else {
+		timeZone = timezones.GetLocalTimeZone()
+	}
+
 	// Configure the module instance.
 	// Note, we use an anonymous module name (empty string) here,
 	// for concurrency and performance reasons.
@@ -123,6 +132,7 @@ func (host *wasmHost) GetModuleInstance(ctx context.Context, plugin *plugins.Plu
 		WithSysWalltime().WithSysNanotime().
 		WithRandSource(rand.Reader).
 		WithStdout(wOut).WithStderr(wErr).
+		WithEnv("TZ", timeZone).
 		WithEnv("CLAIMS", jwtClaims)
 
 	// Instantiate the plugin as a module.