+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - test: add tests for AssemblyScript SDK Transform [#659](https://github.com/hypermodeinc/modus/pull/659)
 - fix: improve Go version handling [#660](https://github.com/hypermodeinc/modus/pull/660)
 - fix: update runtime wasm tests [#661](https://github.com/hypermodeinc/modus/pull/661)
-- fix: support TinyGo 0.35.0 [#661](https://github.com/hypermodeinc/modus/pull/662)
+- fix: support TinyGo 0.35.0 [#662](https://github.com/hypermodeinc/modus/pull/662)
+- feat: add APIs for local time and time zones [#663](https://github.com/hypermodeinc/modus/pull/663)
 
 ## 2024-12-13 - Runtime 0.15.0